feat(sdk): Adding options to client calls
Borrowing optional API design from Dave Cheney
https://dave.cheney.net/2014/10/17/functional-options-for-friendly-apis

I haven't seen the documented forms of `metadata` yet, so here is a general prototype for all metadata.
This starts to get a little harry when we also allow consumers to add/modify grpc.CallOptions to requests.

Finally, I also added the `Options` semantics to the BindingHandler logic.
Personally, this feels a little dirty and should probably be revised before moving forward with this.