Introduce interface for topic event subscriber (#661)
* Introduce interface for topic event subscriber

I introduced the TopicEventSubscriber interface to allow for
subscribers to be implemented as structs. This will allow subscribers
to be initialized with any settings or dependencies needed to process
incoming events.

I converted TopicEventHandler to implement TopicEventSubscriber. I
revised TopicRegistrar to store TopicEventSubscribers instead of
TopicEventHandlers.

Resolves #660

Signed-off-by: Michael Collins <mfcollins3@me.com>

* Update go-service documentation

I added examples of how to use the TopicEventSubscriber interface to
create a new subscriber for both HTTP and gRPC services.

Signed-off-by: Michael Collins <mfcollins3@me.com>

---------

Signed-off-by: Michael Collins <mfcollins3@me.com>